requirements.txt
discord.py==0.9.2
gfycat==0.1.3
giphypop==0.2
requests==2.9.1
ws4py==0.3.4
